+++ src/Queries.EntityFrameworkCore.Extensions/MigrationBuilderExtensions.cs
@@ -1,0 +1,23 @@
+﻿using Microsoft.EntityFrameworkCore.Migrations;
+using Microsoft.EntityFrameworkCore.Migrations.Operations;
+using Queries.Core.Builders;
+using Queries.EntityFrameworkCore.Extensions.Operations;
+
+namespace Queries.EntityFrameworkCore.Extensions
+{
+    public static class MigrationBuilderExtensions
+    {
+
+        /// <summary>
+        ///  Adds a <see cref="MigrationOperation"> to create a view.
+        /// </summary>
+        /// <param name="builder"></param>i        /// <param name="query"></param>
+        /// <param name="schema"></param>
+        /// <returns></returns>
+        public static MigrationBuilder CreateView(this MigrationBuilder builder, CreateViewQuery query, string schema = null)
+        {
+            builder.Operations.Add(new CreateViewMigrationOperation(query, schema));
+            return builder;
+        }
+    }
+}