+++ test/Queries.EntityFrameworkCore.Extensions.SqlServer.Tests/CustomSqlServerMigrationGeneratorTests.cs
@@ -1,0 +1,12 @@
+using System;
+using Microsoft.EntityFrameworkCore.Migrations;
+using Queries.Core.Builders;
+using Xunit;
+using static Queries.Core.Builders.Fluent.QueryBuilder;
+
+namespace Queries.EntityFrameworkCore.Extensions.SqlServer.Tests
+{
+    public class CustomSqlServerMigrationGeneratorTests
+    {
+    }
+}