+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Adds `Queries.EntityFrameworkCore.Extensions` to provide additional `MigrationOperation` implementation
+- Adds [`CreateViewMigrationOperation`](./src/Queries.EntiyFrameworkCore.Extensions/Operations/CreateViewMigrationOperation) to create a SQL view.
 
 ## [0.1.0] / 2021-01-08
 - Initial release